+++ tests_system/lobster_online_report/lobster_online_report_test_runner.py
@@ -7,9 +7,7 @@ from lobster.tools.core.online_report.online_report import main
 
 @dataclass
 class CmdArgs:
-    lobster_report: Optional[str] = None
-    repo_root: Optional[str] = None
-    out: Optional[str] = None
+    config: Optional[str] = None
 
     def as_list(self) -> List[str]:
         """Returns the command line arguments as a list"""
@@ -19,11 +17,7 @@ class CmdArgs:
             if value is not None:
                 cmd_args.append(f"{parameter}={value}")
 
-        if self.lobster_report:
-            cmd_args.append(self.lobster_report)
-
-        append_if_string("--repo-root", self.repo_root)
-        append_if_string("--out", self.out)
+        append_if_string("--config", self.config)
         return cmd_args