+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 ### 0.14.4-dev
 
 * Increased the TRLC version to 2.0.2 when running `lobster-trlc` with Bazel.
+* `lobster-online-report`:
+  - Improved support for git submodules.
+  - Now requires a config file as a command line argument.
 
 ### 0.14.3